requirements.txt
Flask==3.1.3
Werkzeug==3.1.8
openpyxl==3.1.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
redis>=5.0.0
alembic>=1.13.0
SQLAlchemy>=2.0.0
